[QOL] Strict compiler warnings, code clean under all of them
- GCC/Clang: conversions, sign conversions, shadowing, old-style casts and more
- Constructor parameters no longer shadow the members they initialize (also MSVC C4458)
- Grid cells accessed through a helper, line clearing rewritten with unsigned indices
- Score easing computed in double without implicit promotions

+++ src/Piece.cpp
@@ -1,5 +1,7 @@
 #include "Piece.h"
 
+#include <cstddef>
+
 namespace {
 
 struct Definition {
@@ -48,7 +50,7 @@ constexpr Tests SRS_I[4][2] = {
 };
 
 const Definition& Def(TypePiece type) {
-    return DEFINITIONS[static_cast<int>(type)];
+    return DEFINITIONS[static_cast<std::size_t>(type)];
 }
 
 int Normaliser(int rotation) {
